Rename diceThrower to dice-thrower
