+++ src/3d/index.ts
@@ -1,2 +1,2 @@
-export { DiceThrower } from "./diceThrower";
+export { DiceThrower } from "./dice-thrower";
 export { type DieOptions } from "./options";
